Disabled sensorAdjustments during rotation
While standing still and rotating the sensorAdjustment loop was too strong for the rotational PD controller so they fought back and fourth causing the mouse to sometimes stand still and wiggle. By disabling the sensorAdjustments while standing still and rotating this does not happn. This is just a quick fix and there probably alot better ways to do it.